requirements.txt
fastapi
uvicorn
pydantic
pydantic-core
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy 
alembic 
email-validator 
pydantic[email] 
python-multipart
slowapi 
redis 
aioredis
